Feat: Update 'LikeButton' to show loading state
Now the Like button will have its loading state and spinner.
This loading spinner will be shown in two situations, at first render, i.e
when the likes are being fetch, and when the user likes a project, i.e if
the user likes it, it will render the spinner until the request its finish and
showing the updated number of likes or an error message if the request fails.

This will certaintly enhance the user experience.