Feat: Integrate projects likes fetching and error handling in 'ProjectsPage'
- Imported 'involvement' from 'services/involvementAPI/involvementAPI.ts'.
- Set a new local state for the projects likes that will be fetched from the API.
- Created the asynchronous function 'getProjectsLikes' that fetches all the
projects likes within a try-catch block. If successful, the likes are set
in the local state. If an error occurs, it's logged, and the local state is
set to an object with 'error' and 'message' properties for error handling in
the UI component 'LikeButton'.
- Passed down the new local state 'projectsLikes' to
'SplideCarousel' => 'SplideCarouselSlide' to reach the 'LikeButton'
component for error handling.